Fix event number filling in hltanalyzer (#377)

+++ HeavyIonsAnalysis/EventAnalysis/plugins/TriggerAnalyzer.cc
@@ -35,7 +35,7 @@ public:
 private:
   TTree* t_;
 
-  int fEvent;
+  unsigned long long fEvent;
   int fLumiBlock;
   int fRun;
   int fBx;
